fixing the python3 build on unixen.
Actually, this worked mostly by removing the python3 special modules.
After that, things started to build.

It has to be checked, if this works on windows and OS X as well.
Also, there is a build error on Debian and qt 5.3.X at the moment.
But at least, it works better than before.